### 2021-10-21 81ed0a3
some readme typos (#104)
