+++ manuscript/paper.bib
@@ -68,6 +68,7 @@ journal={Nature},
 year={2017},
 volume={550},
 pages={61 - 66}
+doi="10.1186/s13059-017-1359-z"
 }
 
 @article{Mallick2017ExperimentalDA,
@@ -112,7 +113,8 @@ journal={Journal of Open Research Software},
 volume={5},
 number={1},
 year={2017},
-publisher={Ubiquity Press}
+publisher={Ubiquity Press},
+doi={10.5334/jors.151}
 }
 
 @inproceedings{ge2018t,
