+++ .github/workflows/downstream.yml
@@ -17,7 +17,7 @@ jobs:
         julia-version: [1]
         os: [ubuntu-latest]
         package:
-          - {user: EcoJulia, repo: BiobakeryUtils.jl, group: BioSequences}
+          - {user: EcoJulia, repo: BiobakeryUtils.jl, group: Microbiome}
     steps:
       - uses: actions/checkout@v2
       - uses: julia-actions/setup-julia@v1
