### 2021-10-22 77183f0
use taxon instead of Taxon for gf constructor
