+++ manuscript/paper.bib
@@ -114,7 +114,7 @@ volume={5},
 number={1},
 year={2017},
 publisher={Ubiquity Press},
-doi={10.5334/jors.151}
+doi="10.5334/jors.151"
 }
 
 @inproceedings{ge2018t,
