+++ manuscript/paper.bib
@@ -67,7 +67,7 @@ author={Jason Lloyd-Price and Anup Mahurkar and Gholamali Rahnavard and Jonathan
 journal={Nature},
 year={2017},
 volume={550},
-pages={61 - 66}
+pages={61 - 66},
 doi="10.1186/s13059-017-1359-z"
 }
 
