### 2021-10-22 bb4d1c7
Various fixes to tutorial (#109)
* use taxon instead of Taxon for gf constructor

* add using SparseArrays

* fix doi ref

* fix doi ref
