+++ README.md
@@ -3,14 +3,14 @@
 ## Latest Release
 
 [![Latest Release](https://img.shields.io/github/release/EcoJulia/Microbiome.jl.svg)](https://github.com/EcoJulia/Microbiome.jl/releases/latest)
-[![Docs stable](https://img.shields.io/badge/docs-stable-blue.svg)](https://biojulia.net/Microbiome.jl/stable/)
+[![Docs stable](https://img.shields.io/badge/docs-stable-blue.svg)](https://docs.ecojulia.org/Microbiome.jl/stable/)
 [![License](https://img.shields.io/badge/license-MIT-green.svg)](https://github.com/EcoJulia/Microbiome.jl/blob/master/LICENSE)
 [![Project Status: Active – The project has reached a stable, usable state and is being actively developed.](https://www.repostatus.org/badges/latest/active.svg)](https://www.repostatus.org/#active)
 
 ## Development Status
 
 ![EcoJulia maintainer: kescobo](https://img.shields.io/badge/EcoJulia%20Maintainer-kescobo-blue.svg)
-[![Docs dev](https://img.shields.io/badge/docs-latest-blue.svg)](https://biojulia.net/Microbiome.jl/latest/)
+[![Docs dev](https://img.shields.io/badge/docs-latest-blue.svg)](https://docs.ecojulia.org/Microbiome.jl/latest/)
 [![CI](https://github.com/EcoJulia/Microbiome.jl/workflows/CI/badge.svg)](https://github.com/EcoJulia/Microbiome.jl/actions?query=workflow%3ACI)
 [![codecov](https://codecov.io/gh/EcoJulia/Microbiome.jl/branch/main/graph/badge.svg)](https://codecov.io/gh/EcoJulia/Microbiome.jl)
 
