### 2021-11-01 #3841 @kescobo
Alright! Finally got the "stable" docs for BiobakeryUtils to build correctly.

@adRn-s and @aguang thanks for agreeing to review! It's up to you of course, but I suggest [starting here](https://docs.ecojulia.org/BiobakeryUtils.jl/stable/gettingstarted/#getting-started) in the docs
