### 2021-11-01 #3841 @adRn-s
Great. I've followed the docs without any problem so far. At least, until the installation of biobakeryutils. But...

Is ok to report issues here? If so:

I'm unable to run the `metaphlan()` function because [this URL](https://cmprod1.cibio.unitn.it/biobakery3/metaphlan_databases/mpa_latest) is 404... or maybe without SSL cert? I tried `http` (without **s**) and only saw a text in my web browser.
