### 2021-11-02 #3876 @whedon
[REVIEW]: Microbiome.jl and BiobakeryUtils.jl - Julia packages for working with microbial community data
**Submitting author:** @kescobo (<a href="http://orcid.org/0000-0003-3200-7533">Bonham, Kevin</a>)
**Repository:** <a href="https://github.com/EcoJulia/Microbiome.jl" target ="_blank">https://github.com/EcoJulia/Microbiome.jl</a>
**Version:** v0.8.2
