### 2021-11-04 #3876 @aguang
I have finished my review and think the package and paper are a nice contribution to the EcoJulia ecosystem. The documentation is overall solid, I've made a few suggestions to the documentation based on my attempts to run some of the functions on my own data, but those are only suggestions rather than necessary revisions. Nice job, I look forward to trying this package out more extensively soon.
