### 2021-11-04 40aa147
add note about julia version; fixes issue #111
