### 2021-11-04 a596a4a
Other improvements to README.md
