### 2021-11-04 32d613f
add dictionary and named tuple constructors; fixes #110
