+++ previews/PR115/search_index.js
@@ -1,0 +1,3 @@
+var documenterSearchIndex = {"docs":
+[{"location":"profiles/","page":"Profiles and Communities","title":"Profiles and Communities","text":"CurrentModule = Microbiome\nDocTestSetup  = quote\n    using Microbiome\n    using Microbiome.SparseArrays\n    using Random\n    Random.seed!(42)\nend","category":"page"},{"location":"profiles/#Working-with-microbial-abundances","page":"Profiles and Communities","title":"Working with microbial abundances","text":"","category":"section"},{"location":"profiles/","page":"Profiles and Communities","title":"Profiles and Communities","text":"The primary type for working with microbial abundances is the CommunityProfile, which is a sparse matrix with MicrobiomeSamples as column indices and features (eg Taxons or GeneFunctions) as row indices.","category":"page"},{"location":"profiles/","page":"Profiles and Communities","title":"Profiles and Communities","text":"For example, let's make a CommunityProfile with 3 samples, 5 species and 5 genera.","category":"page"},{"location":"profiles/","page":"Profiles and Communities","title":"Profiles and Communities","text":"julia> samps = MicrobiomeSample.([\"s1\", \"s2\", \"s3\"])\n3-element Vector{MicrobiomeSample}:\n MicrobiomeSample(\"s1\", {})\n MicrobiomeSample(\"s2\", {})\n MicrobiomeSample(\"s3\", {})\n\njulia> taxa = [[Taxon(\"s$i\", :species) for i in 1:5]; [Taxon(\"g$i\", :genus) for i in 1:5]]\n10-element Vector{Taxon}:\n Taxon(\"s1\", :species)\n Taxon(\"s2\", :species)\n Taxon(\"s3\", :species)\n Taxon(\"s4\", :species)\n Taxon(\"s5\", :species)\n Taxon(\"g1\", :genus)\n Taxon(\"g2\", :genus)\n Taxon(\"g3\", :genus)\n Taxon(\"g4\", :genus)\n Taxon(\"g5\", :genus)\n\njulia> using SparseArrays\n\njulia> mat = spzeros(10, 3); # 10 x 3 matrix filled with zeros\n\njulia> for i in 1:10, j in 1:3 \n           # fill some spots with random values\n           rand() < 0.3 && (mat[i,j] = rand())\n       end\n\njulia> mat\n10×3 SparseMatrixCSC{Float64, Int64} with 10 stored entries:\n  ⋅         ⋅        0.172933\n  ⋅         ⋅         ⋅\n 0.956916   ⋅         ⋅\n 0.422956   ⋅         ⋅\n  ⋅         ⋅         ⋅\n 0.502952   ⋅        0.167169\n  ⋅         ⋅        0.244683\n  ⋅         ⋅        0.143638\n 0.570085  0.249238   ⋅\n  ⋅        0.841643   ⋅\n\njulia> comm = CommunityProfile(mat, taxa, samps)\nCommunityProfile{Float64, Taxon, MicrobiomeSample} with 10 features in 3 samples\n\nFeature names:\ns1, s2, s3...g4, g5\n\nSample names:\ns1, s2, s3","category":"page"},{"location":"profiles/#Accessing-CommunityProfile-contents","page":"Profiles and Communities","title":"Accessing CommunityProfile contents","text":"","category":"section"},{"location":"profiles/","page":"Profiles and Communities","title":"Profiles and Communities","text":"It is easy to get out the underlying abundance matrix, features, and samples using abundances, features, and samples respectively:","category":"page"},{"location":"profiles/","page":"Profiles and Communities","title":"Profiles and Communities","
[…]
