### 2021-11-04 d8b1833
Improve docstrings for feature types; fixes #114
