### 2021-11-04 c584974
docstring refinements
