### 2021-11-05 e65be6f
add vingette for readin in data; fixes #112
