+++ docs/src/profiles.md
@@ -304,6 +304,109 @@ julia> metadata(comm)
  (sample = "s3", subject = "annelle", foo = "baz", other = missing)
 ```
 
+## Importing from tabular data
+
+The `CommunityProfile` type is compatible with `Tables.jl`,
+so should be able to import data from tabular sources
+with the help of some other packages.
+
+For example, using the [`DataFrames.jl`](https://github.com/JuliaData/DataFrames.jl) package
+(which you can install at the REPL with `] add DataFrames`):
+
+```jldoctest
+julia> using DataFrames, Microbiome
+
+julia> df = DataFrame(features = ["gene$i|s__Some_species" for i in 1:5], s1 = 1:2:10, s2 = 0:2:9)
+5×3 DataFrame
+ Row │ features               s1     s2
+     │ String                 Int64  Int64
+─────┼─────────────────────────────────────
+   1 │ gene1|s__Some_species      1      0
+   2 │ gene2|s__Some_species      3      2
+   3 │ gene3|s__Some_species      5      4
+   4 │ gene4|s__Some_species      7      6
+   5 │ gene5|s__Some_species      9      8
+
+julia> gfs = genefunction.(df.features)
+5-element Vector{GeneFunction}:
+ GeneFunction("gene1", Taxon("Some_species", :species))
+ GeneFunction("gene2", Taxon("Some_species", :species))
+ GeneFunction("gene3", Taxon("Some_species", :species))
+ GeneFunction("gene4", Taxon("Some_species", :species))
+ GeneFunction("gene5", Taxon("Some_species", :species))
+
+julia> mss = MicrobiomeSample.(names(df)[2:end])
+2-element Vector{MicrobiomeSample}:
+ MicrobiomeSample("s1", {})
+ MicrobiomeSample("s2", {})
+
+julia> CommunityProfile(Matrix(df[!, 2:end]), gfs, mss)
+CommunityProfile{Int64, GeneFunction, MicrobiomeSample} with 5 features in 2 samples
+
+Feature names:
+gene1, gene2, gene3, gene4, gene5
+
+Sample names:
+s1, s2
+```
+
+Alternatively, with a CSV file, and the [`CSV.jl`](https://github.com/JuliaData/CSV.jl) package:
+
+```@setup csvexample
+open("taxprof.csv", "w") do io
+    print(io, """
+    features,s1,s2
+    s__Escherichia_coli,1,0
+    s__Bifidobacterium_longum,3,2
+    s__Prevotella_copri,5,4
+    """)
+end
+```
+
+```jldoctest csvexample
+julia> # type ';' to get the shell REPL
+
+shell> cat taxprof.csv -p
+features,s1,s2
+s__Escherichia_coli,1,0
+s__Bifidobacterium_longum,3,2
+s__Prevotella_copri,5,4
+
+julia> using CSV, CSV.Tables
+
+julia> tbl = CSV.read("taxprof.csv", Tables.columntable);
+
+julia> txs = taxon.(tbl[1])
+3-element Vector{Taxon}:
+ Taxon("Escherichia_coli", :species)
+ Taxon("Bifidobacterium_longum", :species)
+ Taxon("Prevotella_copri", :species)
+
+julia> mss = [MicrobiomeSample(string(k)) for k in keys(tbl)[2:end]]
+2-element Vector{MicrobiomeSample}:
+ MicrobiomeSample("s1", {})
+ MicrobiomeSample("s2", {})
+
+julia> mat =  hcat([tbl[i] for i in 2:length(tbl)]...)
+3×2 Matrix{Int64}:
+ 1  0
+ 3  2
+ 5  4
+
+julia> CommunityProfile(mat, txs, mss)
[…]
