### 2021-11-05 #3876 @adRn-s
It has been an honor for me to be a reviewer for this software. I have made some suggestions in their respective repositories, and all have positively been taken into account. The core functionality of both packages is in perfect condition. I look forward to seeing these software libraries grow into the whole framework functionality that EcoJulia has to offer.

I would like to take the opportunity to urge that more diversity metrics (alpha and beta) be incorporated. Then, this could include measures of computational efficiency that the Julia language provides as a modern basis. This way, benchmark comparisons to other software suits could be made. Of course, these are not requirements to approve the publication of your scientific article.

For my part, as a JOSS reviewer, I suggest to the editors that they accept this work as is.

Congratulations to all the authors, keep up with the good work!
