### 2021-11-05 de416cd
attempt to resolve Documenter issues
