### 2021-11-05 EcoJulia/Microbiome.jl#117 @kescobo
Use Diversity.jl for alpha and beta diversity
Possibly worth replacing current implementations of Shannon/Simpson alpha diversity, and the `Distances.jl` - based calculations of beta diversity.

cc @adRn-s
