### 2021-11-05 #3876 @kescobo
Thanks @adRn-s and @aguang for the thoughtful reviews and suggestions! I am working on a number of the suggested changes [here](https://github.com/EcoJulia/Microbiome.jl/pull/115).

>I would like to take the opportunity to urge that more diversity metrics (alpha and beta) be incorporated. Then, this could include measures of computational efficiency that the Julia language provides as a modern basis.

There's been some discussion in the `#ecology` stream on zulip about incorporating [`Diversity.jl`](https://github.com/EcoJulia/Diversity.jl), which may end up being an even better solution. I think either pulling Diveristy.jl in as a dependency or just making sure `CommunityProfile`s play nicely, and maybe adding some examples to the doc. I've opened https://github.com/EcoJulia/Microbiome.jl/issues/117 to keep track, and see also https://github.com/EcoJulia/EcoBase.jl/issues/22
