### 2021-11-05 334d75e
solved doctest issues
