### 2021-11-05 64dc7d9
Joss review fixes (#115)
* add note about julia version; fixes issue #111

* Other improvements to README.md

* add dictionary and named tuple constructors; fixes #110

* Improve docstrings for feature types; fixes #114

* docstring refinements

* add vingette for readin in data; fixes #112

* attempt to resolve Documenter issues

* solved doctest issues
