### 2021-11-05 5375db1
dont use rand in docs
