+++ Project.toml
@@ -26,12 +26,3 @@ MultivariateStats = "0.7, 0.8"
 NamedDims = "0.2"
 Tables = "1.2.1"
 julia = "1.6"
-
-[extras]
-Coverage = "a2441757-f6aa-5fb2-8edb-039e3f45d037"
-Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
-Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
-
-[targets]
-coverage = ["Coverage"]
-test = ["Test", "Random"]
