### 2021-11-05 367b0d9
move doctests to documenter build
