+++ test/runtests.jl
@@ -286,6 +286,11 @@ end
     end
     
     @testset "Indexing and Tables integration" begin
+        @test Tables.istable(comm)
+        @test Tables.columnaccess(comm)
+        @test Tables.rowaccess(comm)
+        @test Tables.schema(comm) isa Tables.Schema
+        
         for i in 1:5
             @test abundances(comm[:, "sample$i"]) == mat[:, [i]]
             @test abundances(comm["taxon$i", :]) == mat[[i], :]
@@ -293,6 +298,7 @@ end
 
         @test abundances(comm[r"taxon1", :]) == abundances(comm[["taxon1", "taxon10"], :]) == abundances(comm[[1,10], :])
         @test abundances(comm[:, r"sample[13]"]) == abundances(comm[:,["sample1", "sample3"]]) == abundances(comm[:, [1,3]])
+        @test abundances(comm[r"taxon1", r"sample[13]"]) == abundances(comm[["taxon1", "taxon10"],["sample1", "sample3"]]) == abundances(comm[[1,10], [1,3]])
 
         for (i, col) in enumerate(Tables.columns(comm))
             if i == 1
