### 2021-11-05 7af1a58
Fixes to doctests (#119)
* dont use rand in docs

* remove Random, extras

* move doctests to documenter build

* add codecov to documenter build

* I want green check!

* I want green check! part 2

* I want green check! part 3
