### 2021-11-10 #3876 @will-rowe
Looks good to me.

@kescobo - can you please create a new tagged release and archive it (in zenodo) and then post back here with the version number and archive DOI
