+++ versions.js
@@ -8,5 +8,5 @@ var DOC_VERSIONS = [
   "v0.1",
   "dev",
 ];
-var DOCUMENTER_NEWEST = "v0.8.1";
+var DOCUMENTER_NEWEST = "v0.8.2";
 var DOCUMENTER_STABLE = "stable";
