### 2021-11-12 #3876 @kescobo
@will-rowe

- Microbiome.jl [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.5682344.svg)](https://doi.org/10.5281/zenodo.5682344)
  - v0.8.3
  - doi: 10.5281/zenodo.5682344
- BiobakeryUtils.jl  [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.5682350.svg)](https://doi.org/10.5281/zenodo.5682350)
  - v0.5.4
  - doi: 10.5281/zenodo.5682350
