### 2021-11-14 #3876 @will-rowe
Great - thanks! All looks good to me. The one thing I'd say is that I've not come across a JOSS paper with 2 repositories/archives associated with the paper. I'll tag this submission with the Microbiome.jl archive but the EiC may have some suggestion on how this is usually handled.
